Merged PR 112: Restricting PAT token as an argument for 'az devops login' command
Restricting PAT token as an argument. Since, the fired CLI commands are logged in command history, we don't want to log the PAT token.

**User will no longer be able to add PAT token via login command argument.**

Earlier behavior  still remains the same, i.e  user will be able to add PAT token via user input prompt contents of which won't be echoed on the screen.
 **Now, user can give PAT token in following 2 ways .**
1. stdin [in interactive mode]
![loginCmd.PNG](https://dev.azure.com/AzureDevOpsCliOrg/b28cb3ba-f9c2-4d01-aab4-7de930823eab/_apis/git/repositories/66cdfc4e-a3bd-483c-8aca-586af75aa80f/pullRequests/112/attachments/loginCmd.PNG)

2. via file [in non interactive mode]
Store the PAT in a file and pass it on to login command.
`cat ~/my_password.txt | az devops login --org <URL>`